Add dedicated test class for PassParameterByReferenceQuickFix
