+++ RetailCoder.VBE/Rubberduck.csproj
@@ -349,7 +349,6 @@
     <Compile Include="Inspections\QuickFixes\AddIdentifierToWhiteListQuickFix.cs" />
     <Compile Include="Inspections\QuickFixes\ApplicationWorksheetFunctionQuickFix.cs" />
     <Compile Include="Inspections\QuickFixes\AssignedByValParameterMakeLocalCopyQuickFix.cs" />
-    <Compile Include="Inspections\QuickFixes\AssignedByValParameterQuickFix.cs" />
     <Compile Include="Inspections\QuickFixes\QuickFixHelper.cs" />
     <Compile Include="Inspections\Resources\InspectionsUI.Designer.cs">
       <AutoGen>True</AutoGen>
