+++ RubberduckTests/Inspections/AssignedByValParameterInspectionTests.cs
@@ -6,12 +6,9 @@ using Rubberduck.Inspections;
 using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Inspections.Resources;
 using Rubberduck.Parsing.VBA;
-using Rubberduck.VBEditor.Events;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 using RubberduckTests.Mocks;
 using System.Collections.Generic;
-using System;
-using Rubberduck.Parsing.Symbols;
 using Rubberduck.VBEditor.SafeComWrappers;
 
 namespace RubberduckTests.Inspections
@@ -127,156 +124,6 @@ End Sub
 
         [TestMethod]
         [TestCategory("Inspections")]
-        public void AssignedByValParameter_QuickFixWorks()
-        {
-
-            string inputCode =
-@"Public Sub Foo(Optional ByVal barByVal As String = ""XYZ"")
-    Let barByVal = ""test""
-End Sub";
-            string expectedCode =
-@"Public Sub Foo(Optional ByRef barByVal As String = ""XYZ"")
-    Let barByVal = ""test""
-End Sub";
-
-            var quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            //check when ByVal argument is one of several parameters
-            inputCode =
-@"Public Sub Foo(ByRef firstArg As Long, Optional ByVal barByVal As String = """", secondArg as Double)
-    Let barByVal = ""test""
-End Sub";
-            expectedCode =
-@"Public Sub Foo(ByRef firstArg As Long, Optional ByRef barByVal As String = """", secondArg as Double)
-    Let barByVal = ""test""
-End Sub";
-
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            inputCode =
-@"
-Private Sub Foo(Optional ByVal  _
-    bar _
-    As _
-    Long = 4, _
-    ByVal _
-    barTwo _
-    As _
-    Long)
-bar = 42
-End Sub
-"
-;
-            expectedCode =
-@"
-Private Sub Foo(Optional ByRef  _
-    bar _
-    As _
-    Long = 4, _
-    ByVal _
-    barTwo _
-    As _
-    Long)
-bar = 42
-End Sub
-"
-;
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            inputCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal _xByValbar As Long,  ByVal _
-    barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-            expectedCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal _xByValbar As Long,  ByRef _
-    barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-            
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            inputCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal barTwoon As Long,  ByVal _
-    barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-            expectedCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal barTwoon As Long,  ByRef _
-    barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            inputCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal barTwoon As Long,  ByVal barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-            expectedCode =
-@"Private Sub Foo(ByVal barByVal As Long, ByVal barTwoon As Long,  ByRef barTwo _
-    As _
-    Long)
-barTwo = 42
-End Sub
-";
-
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-            inputCode =
-@"Sub DoSomething(_
-    ByVal foo As Long, _
-    ByRef _
-        bar, _
-    ByRef barbecue _
-                    )
-    foo = 4
-    bar = barbecue * _
-               bar + foo / barbecue
-End Sub
-";
-
-            expectedCode =
-@"Sub DoSomething(_
-    ByRef foo As Long, _
-    ByRef _
-        bar, _
-    ByRef barbecue _
-                    )
-    foo = 4
-    bar = barbecue * _
-               bar + foo / barbecue
-End Sub
-";
-            quickFixResult = ApplyPassParameterByReferenceQuickFixToCodeFragment(inputCode);
-            Assert.AreEqual(expectedCode, quickFixResult);
-
-        }
-
-
-        [TestMethod]
-        [TestCategory("Inspections")]
         public void AssignedByValParameter_IgnoreQuickFixWorks()
         {
             const string inputCode =
@@ -312,16 +159,11 @@ End Sub";
         }
 
 
-        private string ApplyPassParameterByReferenceQuickFixToCodeFragment(string inputCode)
+        private void AssertVbaFragmentYieldsExpectedInspectionResultCount(string inputCode, int expectedCount)
         {
-            var vbe = BuildMockVBEStandardModuleForVBAFragment(inputCode);
-            var inspectionResults = GetInspectionResults(vbe);
-
-            inspectionResults.First().QuickFixes.Single(s => s is PassParameterByReferenceQuickFix).Fix();
-
-            return GetModuleContent(vbe);
+            var inspectionResults = GetInspectionResults(inputCode);
+            Assert.AreEqual(expectedCount, inspectionResults.Count());
         }
-
 
         private string ApplyIgnoreOnceQuickFixToCodeFragment(string inputCode)
         {
@@ -348,15 +190,12 @@ End Sub";
 
         private IEnumerable<Rubberduck.Inspections.Abstract.InspectionResultBase> GetInspectionResults(Mock<IVBE> vbe)
         {
-            var parser = GetMockParseCoordinator(vbe);
+            var parser = MockParser.Create(vbe.Object, new RubberduckParserState(vbe.Object));
+            parser.Parse(new CancellationTokenSource());
+            if (parser.State.Status >= ParserState.Error) { Assert.Inconclusive("Parser Error"); }
+
             var inspection = new AssignedByValParameterInspection(parser.State);
             return inspection.GetInspectionResults();
-        }
-
-        private void AssertVbaFragmentYieldsExpectedInspectionResultCount(string inputCode, int expectedCount)
-        {
-            var inspectionResults = GetInspectionResults(inputCode);
-            Assert.AreEqual(expectedCount, inspectionResults.Count());
         }
 
         private Mock<IVBE> BuildMockVBEStandardModuleForVBAFragment(string inputCode)
@@ -365,14 +204,6 @@ End Sub";
             IVBComponent component;
             return builder.BuildFromSingleStandardModule(inputCode, out component);
 
-        }
-        private ParseCoordinator GetMockParseCoordinator(Mock<IVBE> vbe)
-        {
-            var parser = MockParser.Create(vbe.Object, new RubberduckParserState(vbe.Object));
-
-            parser.Parse(new CancellationTokenSource());
-            if (parser.State.Status >= ParserState.Error) { Assert.Inconclusive("Parser Error"); }
-            return parser;
         }
     }
 }
